+++ compiler/src/iree/compiler/Dialect/Util/Transforms/BUILD
@@ -33,7 +33,6 @@ iree_compiler_cc_library(
         "StripDebugOps.cpp",
         "TestConversion.cpp",
         "TestFloatRangeAnalysis.cpp",
-        "TraceFrameMark.cpp",
     ],
     hdrs = [
         "Passes.h",
