+++ compiler/src/iree/compiler/Dialect/Util/Transforms/CMakeLists.txt
@@ -35,7 +35,6 @@ iree_cc_library(
     "StripDebugOps.cpp"
     "TestConversion.cpp"
     "TestFloatRangeAnalysis.cpp"
-    "TraceFrameMark.cpp"
   DEPS
     ::PassesIncGen
     LLVMSupport
