+++ compiler/src/iree/compiler/Dialect/Util/Transforms/Passes.td
@@ -155,33 +155,4 @@ def TestFloatRangeAnalysis : Pass<"iree-util-test-float-range-analysis", ""> {
   }];
 }
 
-
-//===----------------------------------------------------------------------===//
-// Tracing
-//===----------------------------------------------------------------------===//
-
-def TraceFrameMarkBegin : Pass<"iree-util-trace-frame-mark-begin", "mlir::ModuleOp"> {
-  let summary = "Begins a frame group when using instrumentation tracing.";
-  let options = [
-    Option<"name", "name",
-           "llvm::StringRef", /*default=*/"llvm::StringRef(\"\")",
-           "Name of the frame to mark (always marks the top-level/unnamed frame).">,
-  ];
-  let constructor = [{
-    mlir::iree_compiler::IREE::Util::createTraceFrameMarkBeginPass()
-  }];
-}
-
-def TraceFrameMarkEnd : Pass<"iree-util-trace-frame-mark-end", "mlir::ModuleOp"> {
-  let summary = "Ends a frame group when using instrumentation tracing.";
-  let options = [
-    Option<"name", "name",
-           "llvm::StringRef", /*default=*/"llvm::StringRef(\"\")",
-           "Name of the frame to mark (always marks the top-level/unnamed frame).">,
-  ];
-  let constructor = [{
-    mlir::iree_compiler::IREE::Util::createTraceFrameMarkEndPass()
-  }];
-}
-
 #endif  // IREE_DIALECT_UTIL_PASSES
