+++ compiler/src/iree/compiler/Pipelines/BUILD
@@ -53,7 +53,7 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Modules/HAL/Inline/Transforms",
         "//compiler/src/iree/compiler/Modules/HAL/Loader/Transforms",
         "//compiler/src/iree/compiler/Preprocessing:Passes",
-        "//runtime/src/iree/base:tracing",
+        "//compiler/src/iree/compiler/Utils",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:Pass",
